Proper json serialization in search CLI.
Closes #211
