+++ quickwit-search/Cargo.toml
@@ -17,6 +17,8 @@ thiserror = "1"
 tonic = '0.4.3'
 tokio-stream = '0.1.6'
 tracing = "0.1"
+serde_json = "1"
+serde = { version = "1.0", features = ["derive"] }
 
 [dependencies.tantivy]
 git = 'https://github.com/quickwit-inc/tantivy'
